Renamer/requirements.txt
# Core dependencies for the Spec Header Date Updater
python-docx>=0.8.11
pywin32>=305
psutil>=5.9.0

# GUI framework
PySide6>=6.5.0

# Firebase integration
firebase-admin>=6.2.0
Pyrebase4>=4.7.1

# Utilities
python-dotenv>=1.0.0
